% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_models.R
\name{print.cdr_model}
\alias{print.cdr_model}
\title{Print method for cdr_model}
\usage{
\method{print}{cdr_model}(x, ...)
}
\arguments{
\item{x}{A \code{cdr_model} object.}

\item{...}{Ignored.}
}
\description{
Print method for cdr_model
}
