% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDREGM.R
\docType{package}
\name{CDREGM-package}
\alias{CDREGM}
\alias{CDREGM-package}
\title{CDREGM: Capitalism, Democracy and Rule of Law Economic Growth Models}
\description{
Implements economic growth models within the Capitalism, Democracy, and Rule of Law (CDR) framework. Provides tools for estimating, comparing, and visualising the joint effects of institutional quality, democratic governance, and property-rights protection on long-run economic growth. Supports panel-data estimation, robustness diagnostics, and replication workflows common in comparative political economy research.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Gypsaman/CDREGM}
  \item Report bugs at \url{https://github.com/Gypsaman/CDREGM/issues}
}

}
\author{
\strong{Maintainer}: Cesar Garcia \email{gypsaman@gmail.com}

Authors:
\itemize{
  \item Cesar Garcia \email{gypsaman@gmail.com}
  \item Felipe Llaugel
}

}
\keyword{internal}
